+++ auto_route_generator/lib/builder.dart
@@ -4,8 +4,6 @@ import 'package:source_gen/source_gen.dart';
 import 'auto_route_generator.dart';
 
 Builder autoRouteGenerator(BuilderOptions options) {
-  return LibraryBuilder(
-    AutoRouteGenerator(),
-    generatedExtension: ".auto_route.json",
-  );
+  // gr stands for generated router.
+  return LibraryBuilder(AutoRouteGenerator(), generatedExtension: ".gr.dart");
 }