+++ auto_route_generator/lib/src/route_parameter.dart
@@ -1,5 +1,6 @@
 import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/type.dart';
+import 'package:auto_route_generator/utils.dart';
 
 // holds constructor parameter info to be used
 // in generating route parameters.
@@ -7,6 +8,7 @@ class RouteParameter {
   String type;
   String name;
   bool isPositional;
+  bool isRequired;
   String defaultValueCode;
   List<String> imports = [];
 
@@ -16,46 +18,31 @@ class RouteParameter {
     name = parameterElement.name;
     isPositional = parameterElement.isPositional;
     defaultValueCode = parameterElement.defaultValueCode;
+    // ignore: deprecated_member_use
+    isRequired = parameterElement.isRequired;
 
+    // import type
     _addImport(paramType.element.source.uri);
 
-    // import generic types
+    // import generic types recursively
+    _checkForParameterizedTypes(paramType);
+  }
+
+  void _checkForParameterizedTypes(DartType paramType) {
     if (paramType is ParameterizedType) {
       paramType.typeArguments.forEach((type) {
-        if (type.element.source != null) _addImport(type.element.source.uri);
+        _checkForParameterizedTypes(type);
+        if (type.element.source != null) {
+          _addImport(type.element.source.uri);
+        }
       });
     }
   }
 
   void _addImport(Uri uri) {
-    if (uri == null) return;
-    final path = uri.toString();
-
-    // we don't need to import core dart types
-    // or core flutter types
-    if (!path.startsWith("dart:core/") &&
-        !path.startsWith("package:flutter/")) {
-      imports.add("'$path'");
+    final import = getImport(uri);
+    if (import != null) {
+      imports.add(import);
     }
   }
-
-  Iterable<DartType> getGenericTypes(DartType type) {
-    return type is ParameterizedType ? type.typeArguments : const [];
-  }
-
-  RouteParameter.fromJson(Map json) {
-    type = json['type'];
-    name = json['name'];
-    isPositional = json['isPositional'];
-    defaultValueCode = json['defaultValueCode'];
-    if (json['imports'] != null) imports = json['imports'].cast<String>();
-  }
-
-  Map<String, dynamic> toJson() => {
-        "type": type,
-        "name": name,
-        "isPositional": isPositional,
-        "defaultValueCode": defaultValueCode,
-        if (imports != null) "imports": imports,
-      };
 }