+++ auto_route_generator/lib/string_utils.dart
@@ -1,3 +1,0 @@
-String toLowerCamelCase(String s) {
-  return s[0].toLowerCase() + s.substring(1);
-}