+++ example/lib/main.dart
@@ -1,6 +1,5 @@
+import 'package:example/router.gr.dart';
 import 'package:flutter/material.dart';
-
-import 'router.dart';
 
 void main() => runApp(MyApp());
 
@@ -9,7 +8,8 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       onGenerateRoute: Router.onGenerateRoute,
-      initialRoute: Router.initialRoute,
+      initialRoute: Router.homeScreenRoute,
+      navigatorKey: Router.navigatorKey,
     );
   }
 }