+++ example/lib/screens/home_screen.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/material.dart';
+
+import '../router.gr.dart';
+
+class HomeScreen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(),
+      body: Container(
+        child: Center(
+          child: FlatButton(
+            child: Text("Second Screen"),
+            onPressed: () {
+              Router.navigator
+                  .pushNamed(Router.secondScreenRoute, arguments: SecondScreenArguments(title: "Title"));
+            },
+          ),
+        ),
+      ),
+    );
+  }
+}