+++ example/lib/screens/nested/nested_home_page.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+class NestedHomePage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Container(child: Center(child: Text("Home nested Page")));
+  }
+}