+++ example/lib/screens/nested/nested_second_page.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class NestedSecondPage extends StatelessWidget {
+  final String name;
+  final int id;
+  final double price;
+
+  const NestedSecondPage({this.name, this.id, this.price});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(child: Center(child: Text("Nested Page")));
+  }
+}