+++ auto_route/CHANGELOG.md
@@ -1,4 +1,15 @@
-## [0.1.0]
-* initial release.
+## [0.2.0+1]
+* format README file
+## [0.2.0]
+### Breaking Changes!
+* change to using a single config file instead of annotating the actual widget class due to performance issues.
+* add @MaterialRoute(), @CupertinoRoute() and @CustomRoute() annotations
+* remove @AutoRoute() annotation and add @AutoRouter()
+* handle required parameters.
+* add navigating with a global navigator key [without context].
+* support nested navigators.
 ## [0.1.1]
 * code formatting.
+## [0.1.0]
+* initial release.
+