+++ auto_route_generator/CHANGELOG.md
@@ -1,3 +1,11 @@
+## [0.2.0]
+### Breaking Changes!
+* change to using a single config file instead of annotating the actual widget class due to performance issues.
+* add @MaterialRoute(), @CupertinoRoute() and @CustomRoute() annotations
+* remove @AutoRoute() annotation and add @AutoRouter()
+* handle required parameters.
+* add navigating with a global navigator key [without context].
+* support nested navigators.
 ## [0.1.3]
 * fix more dependencies resolving conflict.
 ## [0.1.2]